agent: Extract bitflip encryption into its own class
